requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic>=2.0,<3
pyyaml==6.0.1
openai==1.35.14
requests==2.32.3
python-multipart==0.0.9
httpx==0.27.0
